fix. Reorganização dos pacotes em components menores
